apps/whereami/whereami-frontend/src/requirements.txt
flask
requests
emoji==2.1.0
flask-cors
grpcio
grpcio-reflection
grpcio-health-checking
prometheus-flask-exporter
py-grpc-prometheus
opentelemetry-api
opentelemetry-sdk
opentelemetry-propagator-b3
opentelemetry-propagator-gcp
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests